refactor(internal): move `internal/solution.hpp` to `interface.hpp`
