--- internal/include/internal.hpp
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "internal/solution.hpp"
