+++ problems/0001/test.cpp
@@ -2,7 +2,7 @@
 
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/matchers/catch_matchers_vector.hpp>
-#include <internal.hpp>
+#include <interface.hpp>
 #include <string>
 #include <vector>
