+++ problems/0004/cpp/interface.cpp
@@ -1,4 +1,4 @@
-#include <internal.hpp>
+#include <interface.hpp>
 #include <stack>
 #include <vector>
