+++ problems/0004/test.cpp
@@ -1,7 +1,7 @@
 #include <yaml-cpp/yaml.h>
 
 #include <catch2/catch_test_macros.hpp>
-#include <internal.hpp>
+#include <interface.hpp>
 #include <string>
 #include <vector>
