+++ problems/0005/c/interface.cpp
@@ -1,4 +1,4 @@
-#include <internal.hpp>
+#include <interface.hpp>
 #include <string>
 
 extern "C" {
