+++ problems/0009/cpp/interface.cpp
@@ -1,4 +1,4 @@
-#include <internal.hpp>
+#include <interface.hpp>
 #include <list>
 
 using namespace std;
