+++ problems/0011/cpp/interface.cpp
@@ -1,4 +1,4 @@
-#include <internal.hpp>
+#include <interface.hpp>
 #include <vector>
 
 using namespace std;
